pyproject.toml
[tool.poetry]
name = "astrow"
version = "0.1.0"
description = "A comprehensive astrology calculation API"
authors = ["Govindarajan Panneerselvam <imgovind@live.com>"]
readme = "README.md"
packages = [{include = "josi", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
sqlmodel = "^0.0.22"
alembic = "^1.13.3"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"
pydantic = "~2.9.0"
pydantic-settings = "^2.6.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.243.0"}
geopy = "^2.4.1"
timezonefinder = "^6.5.9"
pytz = "^2024.1"
astropy = "^6.1.0"
pyswisseph = "^2.10.3.2"
python-multipart = "^0.0.20"
httpx = "^0.27.0"
skyfield = "^1.48"
lunarcalendar = "^0.0.9"
redis = "^5.0.0"
hiredis = "^2.3.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dateutil = "^2.9.0"
jinja2 = "^3.1.3"
inflection = "^0.5.1"
click = "^8.1.7"
requests = "^2.32.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
black = "^24.0.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
factory-boy = "^3.3.0"
faker = "^24.0.0"
greenlet = "^3.2.3"
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn josi.main:app --reload"
generate-crud = "code_generator.cli:generate_crud"